dpif-netdev: add parameters to configure auto load balance
ALB_IMPROVEMENT_THRESHOLD and ALB_PMD_LOAD_TRIGGER default values
can be overriden with "pmd-auto-lb-min-improvement"
and "pmd-auto-lb-pmd-trigger".

Default values may not be suitable for all use cases, and we may want to
experiment a more (or less) aggressive rebalance, either on the pmd load
trigger or on the minimum variance improvement.
(ie if the new queue assignation will be applied or discarded).

$ ovs-vsctl set open_vswitch . other_config:pmd-auto-lb-pmd-trigger=70
$ ovs-vsctl set open_vswitch . other_config:pmd-auto-lb-min-improvement=20

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>